Django-rest-api/requirements.txt
Django==2.1
psycopg2-binary
djangorestframework
